--- project/junbeom-onboarding/src/main/java/com/fastcampus/survey/common/exception/AbstractGlobalExceptionHandler.java
@@ -1,24 +1,0 @@
-package com.fastcampus.survey.common.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-import java.time.LocalDateTime;
-
-@ControllerAdvice
-public abstract class AbstractGlobalExceptionHandler {
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<ErrorResponse> handleGenericException(Exception ex) {
-        ErrorResponse response = new ErrorResponse(
-                ExceptionConstants.INTERNAL_ERROR,
-                "Internal Server Error",
-                ExceptionConstants.INTERNAL_MESSAGE,
-                LocalDateTime.now().toString()
-        );
-        return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-}
