--- project/junbeom-onboarding/src/main/java/com/fastcampus/survey/common/exception/ErrorResponse.java
@@ -1,14 +1,0 @@
-package com.fastcampus.survey.common.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class ErrorResponse {
-    private int status;
-    private String error;
-    private String message;
-    private String timestamp;
-}
-
