--- project/junbeom-onboarding/src/main/java/com/fastcampus/survey/questionary/adapter/in/mapper/SurveyFormJpaMapper.java
@@ -1,1 +1,0 @@
-
