--- project/junbeom-onboarding/src/main/java/com/fastcampus/survey/questionary/application/service/exception/ContentValidationException.java
@@ -1,9 +1,0 @@
-package com.fastcampus.survey.questionary.application.service.exception;
-
-import jakarta.validation.ValidationException;
-
-public class ContentValidationException extends ValidationException {
-    public ContentValidationException(String field, String message) {
-        super(String.format("%s: %s", field, message));
-    }
-}
