+++ project/junbeom-onboarding/src/main/java/fastcampus_inner_circle/junbeom_onboarding/survey/common/exception/ErrorResponse.java
@@ -1,0 +1,14 @@
+package fastcampus_inner_circle.junbeom_onboarding.survey.common.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class ErrorResponse {
+    private int status;
+    private String error;
+    private String message;
+    private String timestamp;
+}
+
