+++ project/junbeom-onboarding/src/test/java/fastcampus_inner_circle/junbeom_onboarding/JunbeomOnboardingApplicationTests.java
@@ -3,6 +3,7 @@ package fastcampus_inner_circle.junbeom_onboarding;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
+
 @SpringBootTest
 class JunbeomOnboardingApplicationTests {
