+++ project/junbeom-onboarding/src/test/resources/application.properties
@@ -1,12 +1,11 @@
-spring.application.name=junbeom-onboarding
-
 spring.datasource.driver-class-name=org.h2.Driver
+#spring.datasource.url=jdbc:h2:mem:testdb;MODE=MySQL
 spring.datasource.url=jdbc:h2:tcp://localhost/~/fastcampus
 spring.datasource.username=sa
 spring.datasource.password=
 
 spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
-spring.jpa.hibernate.ddl-auto=create-drop
+spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
